test fixes post merge issues. 3 failing still
